Branch TezosClient to inherit from an abstract network client
